+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/contract/payment/PaymentContract.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.dto.order.OrderDto;
 import ru.yandex.practicum.dto.payment.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentContract {
@@ -14,13 +15,13 @@ public interface PaymentContract {
     PaymentDto createPayment(@Validated @RequestBody OrderDto orderDto);
 
     @PostMapping("/totalCost")
-    double calculateTotalCost(@Validated @RequestBody OrderDto orderDto);
+    BigDecimal calculateTotalCost(@Validated @RequestBody OrderDto orderDto);
 
     @PostMapping("/refund")
     void refundPayment(@Validated @RequestBody UUID orderId);
 
     @PostMapping("/productCost")
-    double calculateProductCost(@Validated @RequestBody OrderDto orderDto);
+    BigDecimal calculateProductCost(@Validated @RequestBody OrderDto orderDto);
 
     @PostMapping("/failed")
     void failedPayment(@Validated @RequestBody UUID orderId);
