+++ commerce/payment/src/main/java/ru/yandex/practicum/payment/controller/PaymentController.java
@@ -4,7 +4,6 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.yandex.practicum.contract.payment.PaymentContract;
@@ -12,6 +11,7 @@ import ru.yandex.practicum.dto.order.OrderDto;
 import ru.yandex.practicum.dto.payment.PaymentDto;
 import ru.yandex.practicum.payment.service.PaymentService;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @RestController
@@ -27,7 +27,7 @@ public class PaymentController implements PaymentContract {
     }
 
     @PostMapping("/totalCost")
-    public double calculateTotalCost(OrderDto orderDto) {
+    public BigDecimal calculateTotalCost(OrderDto orderDto) {
         return paymentService.calculateTotalCost(orderDto);
     }
 
@@ -37,7 +37,7 @@ public class PaymentController implements PaymentContract {
     }
 
     @PostMapping("/productCost")
-    public double calculateProductCost(OrderDto orderDto) {
+    public BigDecimal calculateProductCost(OrderDto orderDto) {
         return paymentService.calculateProductCost(orderDto);
     }
